패킷 통일, 클라이언트 플레이어 상태 조정함수 추가

+++ NGP_TCP_Server/Packet.h
@@ -11,7 +11,7 @@ struct PacketHeader {
 struct SC_PLAYER_MOVE {
 	short	targetNum;
 
-	int		hp;
+	float	hp;
 	bool	isAlive;
 	float	posX;
 	float	posY;
